Attendances/Student Attendances/requirements.txt
# Database
mysql-connector-python>=8.0.26

# HTTP Requests
requests>=2.26.0
urllib3>=1.26.7

# Configuration
configparser>=5.0.2

# Logging
logging>=0.5.1.2

# Date and Time
python-dateutil>=2.8.2
